move corpus tests to their own file
